Grouping flask login logic
Moving authentication/user load logic into authenticator.
